Make A Website With Login And Register
How to Make a Website with Login and Register, How to Create a Website with Login and Register, How to Make Website with Login, How to Create a Website with Login and Password, How to Create a Website Login Page in HTML, How to Make Website Login Page, Website with Login and Registration HTML CSS & Javascript
To create a Website with Login & Registration Form using HTML, CSS, and vanilla JavaScript, follow the given steps line by line:

1.Create a folder. You can name this folder whatever you want, and inside this folder, create the mentioned files.
2.Create an index.html file. The file name must be index and its extension .html
3.Create a style.css file. The file name must be style and its extension .css
4.Create a script.js file. The file name must be script and its extension .js
5.Download the background image and put this image inside the project folder. This is the website background image.
Once you create these files, paste the given codes into the specified files. If you don’t want to do these then scroll down and download all the source code files of the Website with the Login & Registration Form, by clicking on the given download button.